pyproject.toml
[project]
name = "cognition-core"
dynamic = ["version"] # Make version dynamic
description = "Core integration package for building AI agents with CrewAI, providing configuration management, memory systems, and tool integration"
authors = [{ name = "Nestor Colt", email = "nestor.colt@gmail.com" }]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "crewai[tools]>=0.100.1,<1.0.0",
    "langchain-community>=0.0.27",
    "chromadb>=0.4.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0.0",
    "portkey-ai>=0.3.0",
    "watchdog>=3.0.0",
    "colorama>=0.4.6",
    "pyyaml>=6.0.1",
]

[project.urls]
Homepage = "https://github.com/nestorcolt/cognition-core"
Repository = "https://github.com/nestorcolt/cognition-core"
Documentation = "https://github.com/nestorcolt/cognition-core#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.dynamic]
version = { attr = "cognition_core.__version__" }

[tool.hatch.version]
path = "src/cognition_core/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cognition_core"]

[tool.pytest]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["src"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["src"]
markers = [
    "asyncio: mark test as async/await test",
]
